+++ app/helpers/kitten/components_helper.rb
@@ -2,7 +2,7 @@ module Kitten
   module ComponentsHelper
     def render_category(category)
       id = category[:id]
-      title = default_title(id)
+      title = category[:label]
       add_menu_group(id, title)
 
       content_tag(:h2, class: 'k-Row karl-Title', id: id) { title }
@@ -10,13 +10,13 @@ module Kitten
 
     def render_component(component,
                          title: nil,
-                         description: nil)
-      title ||= default_title(component)
+                         hidden_on_s: false)
+      title ||= category[:component]
       add_menu_component(component, title)
       render 'layouts/kitten/component',
              component: component,
              title: title,
-             description: description
+             hidden_on_s: hidden_on_s
     end
 
     def example(title = nil)
@@ -43,12 +43,5 @@ module Kitten
       group[:components] << { component_id: id, component_title: title }
     end
 
-    # Turns a type, group or component into a title.
-    #
-    # Example:
-    #   default_title('foo/bar/le-spam') #=> 'Le spam'
-    def default_title(path)
-      path.split('/').last.gsub('-', ' ').capitalize
-    end
   end
 end