+++ assets/javascripts/kitten/karl/headers/karl-header.js
@@ -7,29 +7,38 @@ import { ButtonIcon } from 'kitten/components/buttons/button-icon'
 import { Header,
          HeaderItems,
          HeaderItem } from 'kitten/components/headers/header'
+import { Grid, GridCol } from 'kitten/components/grid/grid'
 import { HeaderTitles } from 'kitten/components/headers/header-titles'
 
-export const KarlMenuHeader = () => (
-  <Header className="karl-Header karl-Menu__header"
-          fixed>
-    <HeaderItems>
-      <HeaderItem>
-        <ButtonIcon data-karl-menu-toggle="true"
-                    className="karl-Menu__header__button">
-          <MenuIcon className={ classNames('k-ButtonIcon__svg',
-                                           'karl-Menu__header__icon',
-                                           'karl-Menu__header__icon--closed') } />
-          <CrossIcon className={ classNames('k-ButtonIcon__svg',
-                                            'karl-Menu__header__icon',
-                                            'karl-Menu__header__icon--opened') } />
-        </ButtonIcon>
-      </HeaderItem>
-      <HeaderItem>
-        <KissKissBankBankLogoIcon className={ classNames('k-Header__logo',
-                                                         'karl-Menu__header__logo') } />
-      </HeaderItem>
-      <HeaderTitles title="Styleguide"
-                    className="karl-Menu__header__title" />
-    </HeaderItems>
-  </Header>
+export const KarlMenuHeader = props => (
+  <Grid>
+    <GridCol col="12">
+      <Header className="karl-Header karl-Menu__header"
+              fixed>
+        <HeaderItems>
+          <HeaderItem>
+            <ButtonIcon onClick={ props.onButtonClick }
+                        className="karl-Menu__header__button">
+              <MenuIcon className={ classNames('k-ButtonIcon__svg',
+                                               'karl-Menu__header__icon',
+                                               'karl-Menu__header__icon--closed') } />
+              <CrossIcon className={ classNames('k-ButtonIcon__svg',
+                                                'karl-Menu__header__icon',
+                                                'karl-Menu__header__icon--opened') } />
+            </ButtonIcon>
+          </HeaderItem>
+          <HeaderItem>
+            <KissKissBankBankLogoIcon className={ classNames('k-Header__logo',
+                                                             'karl-Menu__header__logo') } />
+          </HeaderItem>
+          <HeaderTitles title="Styleguide"
+                        className="karl-Menu__header__title" />
+        </HeaderItems>
+      </Header>
+    </GridCol>
+  </Grid>
 )
+
+KarlMenuHeader.defaultProps = {
+  onButtonClick: () => {},
+}