+++ CHANGELOG.md
@@ -4,6 +4,99 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+## [12.1.0] - 2017-06-08
+
+Feature:
+- Create `ResponsiveVideoIframeContainer` component for responsive video iframe.
+  The component is used with embedly generated iframe.
+
+Fixies:
+- Remove `color-error` label on `RadioButton` and `Checkbox` components.
+- Add `box-shadow: none` property on `TextInput`.
+- Feature:
+- Fix error with `Object.assign` in `RichTextInput` component to prevent
+  error in React on Rails *(`Object function […] has no method 'assign'`)*.
+
+## [12.0.0] - 2017-05-30
+
+Features:
+- Add new `ProjectCreatorCard` component.
+- Add new `ClockIcon` component.
+- Add new `RightArrowIcon` component.
+- Add config & toolbar to default props `RichTextInput` component.
+- Fix: Fix display element as a block element (`ButtonImageWithText`,
+    `ButtonImage`, `ButtonWithBadge`, `IconBadge`).
+
+Fixes:
+- Remove `React.PropTypes` and use `PropTypes` from
+  [prop-types](https://www.npmjs.com/package/prop-types) library instead.
+- Update deprecated
+  [react-prop-types](https://github.com/react-bootstrap/react-prop-types)
+  library with
+  [prop-types-extra](https://github.com/react-bootstrap/prop-types-extra).
+
+## [11.11.0] - 2017-05-22
+
+Fix:
+- Change `TagButton` background.
+
+## [11.10.0] - 2017-05-19
+
+Features:
+- Add `BulletList` component.
+- Add `big` modifier to `.k-BulletList`.
+
+## [11.9.0] - 2017-05-18
+
+Feature:
+- Add background color utilities.
+
+Fixes:
+- Remove `margin-top: 50px` on `k-Separator`.
+- Add missing onChange prop on `RichTextInput` component.
+
+## [11.8.1] - 2017-05-17
+
+Fix:
+- Prefer `npm` to `yarn` in the postinstall script.
+
+## [11.8.0] - 2017-05-17
+
+Feature:
+- Add `RichTextInput` component.
+  Add '<script src="//cdn.ckeditor.com/4.6.2/standard/ckeditor.js"></script>'
+  to your pages.
+
+## [11.7.0] - 2017-05-15
+
+Feature:
+- Create `CheckboxSet` & `RadioButtonSet` components.
+
+## [11.6.0] - 2017-05-15
+
+Features:
+- Add tooltip in `ProjectCard`.
+- Add new `MasterCardIcon` and `CbIcon` components.
+- Add `.k-TagButton__icon` class to adds icons to tag buttons.
+
+## [11.5.0] - 2017-05-09
+
+Feature:
+- Add responsive image utility.
+
+Fix:
+- Fix the tiny option on Select.
+
+## [11.4.0] - 2017-05-04
+
+Feature:
+- Add status in `ProjectCard`.
+
+Fixes:
+- Fix `LegoGrid` grid.
+- Fix tag `Title` in `ProjectCardComponent`.
+- Fix test launch.
+
 ## [11.3.0] - 2017-05-02
 
 Features: